feat: add handleHomeChange module in utils
